// packages/deployment/types/base-types.ts
import { Address } from 'viem'

export enum SupportedNetworks {
  MAINNET = 'mainnet',
  BASE = 'base',
  ARBITRUM = 'arbitrum',
  SONIC = 'sonic',
  HYPERLIQUID = 'hyperliquid',
}

export enum ArkType {
  AaveV3Ark = 'AaveV3Ark',
  SparkArk = 'SparkArk',
  CompoundV3Ark = 'CompoundV3Ark',
  CrossChainArk = 'CrossChainArk',
  ERC4626Ark = 'ERC4626Ark',
  MorphoArk = 'MorphoArk',
  MorphoVaultArk = 'MorphoVaultArk',
  PendleLPArk = 'PendleLPArk',
  PendlePTArk = 'PendlePTArk',
  PendlePtOracleArk = 'PendlePtOracleArk',
  SkyUsdsArk = 'SkyUsdsArk',
  SkyUsdsPsm3Ark = 'SkyUsdsPsm3Ark',
  MoonwellArk = 'MoonwellArk',
  SyrupArk = 'SyrupArk',
  SkyRewardsArk = 'SkyRewardsArk',
  SiloArk = 'SiloArk',
  SiloArkV2 = 'SiloArkV2',
  SiloManagedVaultArk = 'SiloManagedVaultArk',
  OriginETHArk = 'OriginETHArk',
  ArmArk = 'ArmArk',
  FluidLiteArk = 'FluidLiteArk',
  AeraArk = 'AeraArk',
  StargateV2PoolArk = 'StargateV2PoolArk',
  SiUSDArk = 'SiUSDArk',
  FluidFTokenArk = 'FluidFTokenArk',
  PsmLiteERC4626Ark = 'PsmLiteERC4626Ark',
  Psm3ERC4626Ark = 'Psm3ERC4626Ark',
  HyperlendArk = 'HyperlendArk',
  HypurrArk = 'HypurrArk',
  WisdomTreeArk = 'WisdomTreeArk',
  MorphoV2VaultArk = 'MorphoV2VaultArk',
  MapleInstitutionalArk = 'MapleInstitutionalArk',
  UpshiftArk = 'UpshiftArk',
  OriginUSDArk = 'OriginUSDArk',
}

export const arkTypes = [
  { title: 'AaveV3Ark', value: ArkType.AaveV3Ark },
  { title: 'SparkArk', value: ArkType.SparkArk },
  { title: 'MorphoArk', value: ArkType.MorphoArk },
  { title: 'MorphoVaultArk', value: ArkType.MorphoVaultArk },
  { title: 'CompoundV3Ark', value: ArkType.CompoundV3Ark },
  { title: 'CrossChainArk', value: ArkType.CrossChainArk },
  { title: 'ERC4626Ark', value: ArkType.ERC4626Ark },
  { title: 'SkyUsdsArk', value: ArkType.SkyUsdsArk },
  { title: 'SkyUsdsPsm3Ark', value: ArkType.SkyUsdsPsm3Ark },
  { title: 'PendleLPArk', value: ArkType.PendleLPArk },
  { title: 'PendlePTArk', value: ArkType.PendlePTArk },
  { title: 'PendlePtOracleArk', value: ArkType.PendlePtOracleArk },
  { title: 'MoonwellArk', value: ArkType.MoonwellArk },
  { title: 'SyrupArk', value: ArkType.SyrupArk },
  { title: 'SkyRewardsArk', value: ArkType.SkyRewardsArk },
  { title: 'SiloArk', value: ArkType.SiloArk },
  { title: 'SiloManagedVaultArk', value: ArkType.SiloManagedVaultArk },
  { title: 'OriginETHArk', value: ArkType.OriginETHArk },
  { title: 'ArmArk', value: ArkType.ArmArk },
  { title: 'FluidLiteArk', value: ArkType.FluidLiteArk },
  { title: 'AeraArk', value: ArkType.AeraArk },
  { title: 'StargateV2PoolArk', value: ArkType.StargateV2PoolArk },
  { title: 'SiUSDArk', value: ArkType.SiUSDArk },
  { title: 'FluidFTokenArk', value: ArkType.FluidFTokenArk },
  { title: 'PsmLiteERC4626Ark', value: ArkType.PsmLiteERC4626Ark },
  { title: 'Psm3ERC4626Ark', value: ArkType.Psm3ERC4626Ark },
  { title: 'HyperlendArk', value: ArkType.HyperlendArk },
  { title: 'HypurrArk', value: ArkType.HypurrArk },
  { title: 'WisdomTreeArk', value: ArkType.WisdomTreeArk },
  { title: 'MapleInstitutionalArk', value: ArkType.MapleInstitutionalArk },
  { title: 'UpshiftArk', value: ArkType.UpshiftArk },
  { title: 'OriginUSDArk', value: ArkType.OriginUSDArk },
]

export enum Token {
  USDC = 'usdc',
  DAI = 'dai',
  USDT = 'usdt',
  USDE = 'usde',
  USDCE = 'usdce',
  USDS = 'usds',
  STAKED_USDS = 'stakedUsds',
  WETH = 'weth',
  STETH = 'steth',
  EURC = 'eurc',
  SEAM = 'seam',
  REUL = 'reul',
  WELL = 'well',
  WS = 'ws',
  GEAR = 'gear',
  MORPHO = 'morpho',
  SYRUP = 'syrup',
  SILO = 'silo',
  SKY = 'sky',
  XSILO = 'xsilo',
  ASONW = 'asonw',
  WRAPPED_NATIVE = 'wrappedNative',
}